20210126_095450

+++ Scripts/vercel-build.sh
@@ -1,5 +1,5 @@
 #! /bin/bash
-# Time-stamp: "2021-01-25 18:29:52 queinnec"
+# Time-stamp: "2021-01-26 10:54:08 queinnec"
 
 # Build the P server on Vercel.
 # api/p.js should already exist to be taken into account.
@@ -93,8 +93,30 @@ else
     unset FKEY_PRIVATE
 fi
 
-# Preserve this server.js since it is required for sapper export:
-mv src/server.js src/server.js.ORG
+echo "*** Building the p.codegradx.org static server..."
+# export needs the original server.js with a listen method in order to
+# crawl the site after generation.
+if grep -q listen.PORT < src/server.js
+then :
+else
+    echo "Bad version of src/server.js"
+    exit 2
+fi
+
+npm run export
+showls `pwd`/
+showls __sapper__/
+showls __sapper__/export/
+
+if [ ! -f __sapper__/export/index.html ]
+then
+    echo "!!!!! Crawling phase probably missing!" >&2
+    exit 1
+fi
+# Make sure that export is the 'output directory' in the project settings:
+mv __sapper__/export .
+
+# Update src/server.js in order to be invoked as a serverless function.
 cat > src/server.js <<EOF
 import sirv from 'sirv';
 import polka from 'polka';
@@ -102,12 +124,11 @@ import compression from 'compression';
 import * as sapper from '@sapper/server';
 import { CodeGradX } from 'codegradx';
 const serverless = require('serverless-http');
+const fs = require('fs');
+const path = require('path');
 
 const { PORT, NODE_ENV } = process.env;
 const dev = NODE_ENV === 'development';
-
-const fs = require('fs');
-const path = require('path');
 
 function showDir (dir = '.') {
   try {
@@ -121,17 +142,11 @@ function showDir (dir = '.') {
 showDir('.');
 showDir('./api');
 showDir('./static');
+showDir('./export');
 showDir('./__sapper__');
 showDir('./__sapper__/build');
-showDir('./__sapper__/export');
-showDir('./__sapper__/export/static');
-showDir('./__sapper__/export/__sapper__');
-showDir('./__sapper__/export/__sapper__/build');
 
 const staticDir = path.join(process.cwd(), 'static');
-
-process.chdir('./__sapper__/export');
-showDir(process.cwd());
 
 let handler = undefined;
 try {
@@ -141,12 +156,6 @@ try {
                 sirv(staticDir, { dev }),
                 sapper.middleware()
         );
-  if ( dev ) {
-     console.log('listening on ' + PORT);
-     server.listen(PORT, err => {
-          if (err) console.log('error', {err});
-     });
-  }
   handler = serverless(server);
 } catch (exc) {
   console.error({exc});
@@ -166,51 +175,7 @@ EOF
 
 echo "*** Building the p.codegradx.org dynamic server..."
 npm run build
-showls `pwd`/
-showls __sapper__/
-#showls __sapper__/export/
-showls __sapper__/build/
-$DEBUG && tail -20 __sapper__/build/server/server.js
-
-false && \
-( cd __sapper__/build/ && sed -i.bak \
-      -e 's@__sapper__/build@./__sapper__/build@' \
-      server/server.js && \
-  rm server/server.js.bak && \
-  cp -p server/server.js ../../src/server.js.SAV )
-cp __sapper__/build/server/server.js src/server.js.SAV 
-# Preserve that server.js it will be destroyed by sapper export.
-
-echo "*** Building the p.codegradx.org static server..."
-# Restaure this server.js since it is required by export when crawling the site:
-mv src/server.js.ORG src/server.js
-npm run export
-showls `pwd`/
-showls __sapper__/
-showls __sapper__/export/
-
-if [ ! -f __sapper__/export/index.html ]
-then
-    echo "!!!!! Crawling phase probably missing!" >&2
-    exit 1
-fi
-
-mkdir -p __sapper__/export/static/
-#cp -p __sapper__/export/*.{ico,png,css,json,html,js} __sapper__/export/static/
-cp -rp static __sapper__/export/
-showls __sapper__/
-showls __sapper__/export/
-
-mkdir -p __sapper__/export/__sapper__/
-mv src/server.js.SAV __sapper__/build/server/server.js
-mv __sapper__/build __sapper__/export/__sapper__/
-
-# NOTA: the content of p.js cannot be changed at build-time:
-#mkdir -p api
-#cp -p __sapper__/export/sapper/build/server/server.js api/
-#showls api/
-
-showls -R __sapper__/export/
+cp __sapper__/build/server/server.js api/p.js
 
 echo "*** end of vercel-build.sh"
